Move precompiles->handlers
