+++ stdlib/tests/main.rs
@@ -31,6 +31,6 @@ mod crypto;
 mod mast_forest_merge;
 mod math;
 mod mem;
-mod precompiles;
+mod handlers;
 mod sys;
 mod word;
